linkedin-posts-python/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
jinja2>=3.1,<4
python-multipart>=0.0.9,<1
python-dotenv>=1.0,<2
playwright>=1.43,<2
